Update Haversine service test
